requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==5.1.5
django-cors-headers==4.9.0
django-filter==24.3
django-simpleui==2024.11.15
django-templated-mail==1.1.1
djangorestframework==3.16.1
djoser==2.2.3
idna==3.11
mysqlclient==2.2.7
pillow==12.1.0
python-decouple==3.8
requests==2.32.5
schedule==1.2.2
google-play-scraper==1.2.7
feedparser==6.0.11
lxml==5.3.0
beautifulsoup4==4.14.3
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
# Optional local TTS engine (used by /api/tts/coqui/synthesize/)
# Coqui TTS currently supports Python < 3.12.
coqui-tts>=0.27.5; python_version < "3.12"
pypinyin>=0.55.0
